requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==4.0.1
attrs==23.2.0
browserstack-local==1.2.7
browserstack_sdk==1.20.2
certifi==2024.7.4
charset-normalizer==3.3.2
dukpy==0.4.0
execnet==2.1.1
filelock==3.15.4
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.7
iniconfig==2.0.0
mutf8==1.0.6
outcome==1.3.0.post0
packaging==24.1
percy-appium-app==2.0.5
pluggy==1.5.0
psutil==6.0.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-SystemConfiguration==10.3.1
PyPAC==0.16.4
PySocks==1.7.1
pytest==8.2.2
pytest-xdist==3.6.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
selenium==4.22.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
tldextract==5.1.2
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
